sorced out the pdf reader and began to implement navigation
The code for starting and managing the pdf reader has been completely outsourced to PDFManager. At the moment this class is not finished as it lacks error prevention and good module design. When finished there should only be 4 functions being accessible from the module: change the current PDF, render this PDF, render next page, render previous page.
Also a little navigation has been added to the reader but needs to be layoutet in future commits.