implemented basic part of comment section
The first version of the comment section has been added. It is now possible to write comments for the whole pdf document. The partitioning per page is still a thing to do, as well as the css and js for adding audio and videofiles. At the moment it is only possible to write text as comments. And also the comments are currently not saved to a database so they get lost after page reload.